Add gemini as a partner provider
